refactor(masking): move masking implementations to masking crate
